Clean up clippy warnings in agc-core
Resolves 2 deny-level errors (approx_constant) and 41 unique warnings
across 16 files. No behavioural changes; all 405 tests pass and the
bare-metal thumbv7em-none-eabihf build still compiles.

- needless_range_loop: convert index-based loops to enumerate/zip/iter
- single_match: replace `match … { Variant => …, _ => {} }` with `if`
- len_zero: prefer `!is_empty()` over `len() >= 1`
- identity_op: drop redundant `0xFFFF & …` masks; rewrite tautological
  16-bit-fit assertion as a typed-binding documentation
- manual_range_contains: use `(a..=b).contains(&x)`
- doc_overindented_list_items / doc_lazy_continuation: reflow doc lists
- field_reassign_with_default: use struct-init shorthand
- new_without_default: add `Default` for `RestartProtection`, `AgcState`
- derivable_impls: replace manual `Default for VnPhase` with `#[derive]`
- type_complexity: introduce `pub type TaskFn = fn(&mut AgcState)`
- items_after_test_module: hoist `update_dsky_n43` (p22) and
  `propagate_to_tig` (p31) above their `mod tests` blocks
- assertions_on_constants: move const check into `const _: () = assert!`
- manual_range_patterns: `21 | 22 | 23` → `21..=23`
- approx_constant: replace `3.14` test pollution with `3.5`; replace
  `0.523_6` with `core::f64::consts::FRAC_PI_6`
- let_and_return, unnecessary_cast: trivial cleanups

For the `RESTART_GROUP_TABLE` (`static mut`) iteration in
`fresh_start.rs`, switched to a single explicit borrow via
`&*core::ptr::addr_of!(...)` so the iterator-style loop does not
trigger `static_mut_refs`.

+++ agc-core/src/guidance/maneuver.rs
@@ -183,7 +183,7 @@ pub fn is_burn_complete(state: &BurnState, cutoff_tolerance: f64) -> bool {
 /// - Returns a `Met` representing the estimated absolute cutoff time.
 /// - On the fallback path, returns `state.tig + MAX_BURN_DURATION_CS`.
 pub fn compute_cutoff_time(state: &BurnState, current_met: Met) -> Met {
-    let elapsed_cs = current_met.0.wrapping_sub(state.tig.0) as u32;
+    let elapsed_cs = current_met.0.wrapping_sub(state.tig.0);
     let elapsed_s = elapsed_cs as f64 / 100.0;
 
     if elapsed_s < 0.1 {
